refactor: move getResponseBody function from useNetworkTracker to NetworkProvider, in order to preserve responseBodies state between copmonents
