+++ packages/vscode-extension/src/network/hooks/useNetworkTracker.tsx
@@ -67,7 +67,6 @@ export interface WebSocketMessage {
 export interface NetworkTracker {
   networkLogs: NetworkLog[];
   ws: WebSocket | null;
-  getResponseBody: (networkLog: NetworkLog) => Promise<unknown>;
   clearLogs: () => void;
   toggleNetwork: (isRunning: boolean) => void;
   getSource: (networkLog: NetworkLog) => void;
@@ -76,7 +75,6 @@ export interface NetworkTracker {
 export const networkTrackerInitialState: NetworkTracker = {
   networkLogs: [],
   ws: null,
-  getResponseBody: async () => undefined,
   clearLogs: () => {},
   toggleNetwork: () => {},
   getSource: () => {},
@@ -182,46 +180,6 @@ const useNetworkTracker = (): NetworkTracker => {
     );
   };
 
-  const [responseBodies, setResponseBodies] = useState<Record<string, unknown>>({});
-
-  const getResponseBody = (networkLog: NetworkLog) => {
-    if (responseBodies[networkLog.requestId]) {
-      return Promise.resolve(responseBodies[networkLog.requestId]);
-    }
-
-    const id = Math.random().toString(36).substring(7);
-
-    wsRef.current?.send(
-      JSON.stringify({
-        id,
-        method: "Network.getResponseBody",
-        params: {
-          requestId: networkLog.requestId,
-        },
-      })
-    );
-
-    return new Promise((resolve) => {
-      const listener = (message: MessageEvent) => {
-        try {
-          const parsedMsg = JSON.parse(message.data);
-          if (parsedMsg.id === id) {
-            setResponseBodies((prev) => ({
-              ...prev,
-              [networkLog.requestId]: parsedMsg.result.body,
-            }));
-            resolve(parsedMsg.result.body);
-            wsRef.current?.removeEventListener("message", listener);
-          }
-        } catch (error) {
-          console.error("Error parsing WebSocket message:", error);
-        }
-      };
-
-      wsRef.current?.addEventListener("message", listener);
-    });
-  };
-
   const getSource = (networkLog: NetworkLog) => {
     wsRef.current?.send(
       JSON.stringify({
@@ -236,7 +194,6 @@ const useNetworkTracker = (): NetworkTracker => {
   return {
     networkLogs: networkLogs.filter((log) => log?.request?.url !== undefined),
     ws: wsRef.current,
-    getResponseBody,
     clearLogs,
     toggleNetwork,
     getSource,
