+++ packages/vscode-extension/src/network/providers/NetworkProvider.tsx
@@ -7,6 +7,7 @@ import React, {
   useState,
 } from "react";
 import useNetworkTracker, {
+  NetworkLog,
   NetworkTracker,
   networkTrackerInitialState,
 } from "../hooks/useNetworkTracker";
@@ -20,6 +21,7 @@ interface NetworkProviderProps extends NetworkTracker {
   toggleScrolling: () => void;
   isTimelineVisible: boolean;
   toggleTimelineVisible: () => void;
+  getResponseBody: (networkLog: NetworkLog) => Promise<unknown>;
 }
 
 const NetworkContext = createContext<NetworkProviderProps>({
@@ -31,6 +33,7 @@ const NetworkContext = createContext<NetworkProviderProps>({
   toggleScrolling: () => {},
   isTimelineVisible: true,
   toggleTimelineVisible: () => {},
+  getResponseBody: async () => undefined,
 });
 
 export default function NetworkProvider({ children }: PropsWithChildren) {
@@ -39,15 +42,56 @@ export default function NetworkProvider({ children }: PropsWithChildren) {
   const [isTimelineVisible, toggleTimelineVisible] = useReducer((state) => !state, true);
   const [isScrolling, toggleScrolling] = useReducer((state) => !state, false);
   const [isRecording, setIsRecording] = useState(true);
+  const [responseBodies, setResponseBodies] = useState<Record<string, unknown>>({});
 
   const clearActivity = () => {
     networkTracker.clearLogs();
+    setResponseBodies({});
   };
 
   const toggleRecording = () => {
     setIsRecording((prev) => {
       networkTracker.toggleNetwork(prev);
       return !prev;
+    });
+  };
+
+  const getResponseBody = (networkLog: NetworkLog) => {
+    const ws = networkTracker.ws;
+    if (responseBodies[networkLog.requestId]) {
+      return Promise.resolve(responseBodies[networkLog.requestId]);
+    }
+
+    const id = Math.random().toString(36).substring(7);
+
+    ws?.send(
+      JSON.stringify({
+        id,
+        method: "Network.getResponseBody",
+        params: {
+          requestId: networkLog.requestId,
+        },
+      })
+    );
+
+    return new Promise((resolve) => {
+      const listener = (message: MessageEvent) => {
+        try {
+          const parsedMsg = JSON.parse(message.data);
+          if (parsedMsg.id === id) {
+            setResponseBodies((prev) => ({
+              ...prev,
+              [networkLog.requestId]: parsedMsg.result.body,
+            }));
+            resolve(parsedMsg.result.body);
+            ws?.removeEventListener("message", listener);
+          }
+        } catch (error) {
+          console.error("Error parsing WebSocket message:", error);
+        }
+      };
+
+      ws?.addEventListener("message", listener);
     });
   };
 
@@ -62,6 +106,7 @@ export default function NetworkProvider({ children }: PropsWithChildren) {
       toggleScrolling,
       isTimelineVisible,
       toggleTimelineVisible,
+      getResponseBody,
     };
   }, [isRecording, isScrolling, isTimelineVisible, networkTracker.networkLogs]);
